+++ homeassistant/components/gogogate2/__init__.py
@@ -1,6 +1,8 @@
 """The gogogate2 component."""
-from homeassistant.components.cover import DOMAIN as COVER_DOMAIN
-from homeassistant.components.sensor import DOMAIN as SENSOR_DOMAIN
+import asyncio
+
+from homeassistant.components.cover import DOMAIN as COVER
+from homeassistant.components.sensor import DOMAIN as SENSOR
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_DEVICE
 from homeassistant.core import HomeAssistant
@@ -8,6 +10,8 @@ from homeassistant.exceptions import ConfigEntryNotReady
 
 from .common import get_data_update_coordinator
 from .const import DEVICE_TYPE_GOGOGATE2
+
+PLATFORMS = [COVER, SENSOR]
 
 
 async def async_setup(hass: HomeAssistant, base_config: dict) -> bool:
@@ -35,20 +39,23 @@ async def async_setup_entry(hass: HomeAssistant, config_entry: ConfigEntry) -> b
     if not data_update_coordinator.last_update_success:
         raise ConfigEntryNotReady()
 
-    hass.async_create_task(
-        hass.config_entries.async_forward_entry_setup(config_entry, COVER_DOMAIN)
-    )
-    hass.async_create_task(
-        hass.config_entries.async_forward_entry_setup(config_entry, SENSOR_DOMAIN)
-    )
+    for platform in PLATFORMS:
+        hass.async_create_task(
+            hass.config_entries.async_forward_entry_setup(config_entry, platform)
+        )
 
     return True
 
 
 async def async_unload_entry(hass: HomeAssistant, config_entry: ConfigEntry) -> bool:
     """Unload Gogogate2 config entry."""
-    hass.async_create_task(
-        hass.config_entries.async_forward_entry_unload(config_entry, COVER_DOMAIN)
+    unload_ok = all(
+        await asyncio.gather(
+            *[
+                hass.config_entries.async_forward_entry_unload(config_entry, platform)
+                for platform in PLATFORMS
+            ]
+        )
     )
 
-    return True
+    return unload_ok
