+++ homeassistant/components/gogogate2/sensor.py
@@ -1,20 +1,16 @@
 """Support for Gogogate2 garage Doors."""
 from typing import Callable, List, Optional
 
-from gogogate2_api.common import AbstractDoor, get_configured_doors, get_door_by_id
+from gogogate2_api.common import get_configured_doors
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import DEVICE_CLASS_BATTERY
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity import Entity
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from .common import (
-    DeviceDataUpdateCoordinator,
-    cover_unique_id,
-    get_data_update_coordinator,
-)
-from .const import DOMAIN
+from .common import GoGoGate2Entity, get_data_update_coordinator
+
+SENSOR_ID_WIRED = "WIRE"
 
 
 async def async_setup_entry(
@@ -29,32 +25,13 @@ async def async_setup_entry(
         [
             DoorSensor(config_entry, data_update_coordinator, door)
             for door in get_configured_doors(data_update_coordinator.data)
-            if door.sensorid and door.sensorid != "WIRE"
+            if door.sensorid and door.sensorid != SENSOR_ID_WIRED
         ]
     )
 
 
-class DoorSensor(CoordinatorEntity):
+class DoorSensor(GoGoGate2Entity):
     """Sensor entity for goggate2."""
-
-    def __init__(
-        self,
-        config_entry: ConfigEntry,
-        data_update_coordinator: DeviceDataUpdateCoordinator,
-        door: AbstractDoor,
-    ) -> None:
-        """Initialize the object."""
-        super().__init__(data_update_coordinator)
-        self._config_entry = config_entry
-        self._door = door
-        self._api = data_update_coordinator.api
-        self._unique_id = cover_unique_id(config_entry, door)
-        self._is_available = True
-
-    @property
-    def unique_id(self) -> Optional[str]:
-        """Return a unique ID."""
-        return self._unique_id
 
     @property
     def name(self):
@@ -70,7 +47,7 @@ class DoorSensor(CoordinatorEntity):
     def state(self):
         """Return the state of the entity."""
         door = self._get_door()
-        return door.voltage
+        return door.voltage  # This is a percentage, not an absolute voltage
 
     @property
     def state_attributes(self):
@@ -80,15 +57,3 @@ class DoorSensor(CoordinatorEntity):
         if door.sensorid is not None:
             attrs["sensorid"] = door.door_id
         return attrs
-
-    def _get_door(self) -> AbstractDoor:
-        door = get_door_by_id(self._door.door_id, self.coordinator.data)
-        self._door = door or self._door
-        return self._door
-
-    @property
-    def device_info(self):
-        """Device info for the controller, details are set by the cover entity."""
-        return {
-            "identifiers": {(DOMAIN, self._config_entry.unique_id)},
-        }
